I don't remember what i added, but the Lumia can boot very very minimal version of kernel

+++ include/arch/x86/io.h
@@ -1,8 +1,8 @@
 #ifndef IO_H
 #define IO_H
 #include <typedefs.h>
-void outb(uint32_t port,uint8_t data);
-uint8_t inb(uint32_t port);
+void outb(uint16_t port,uint8_t data);
+uint8_t inb(uint16_t port);
 void outw(uint32_t port,uint16_t data);
 uint16_t inw(uint32_t port);
 void insw(uint16_t port, unsigned char *data, unsigned long size);
