+++ src/arch/x86/io.c
@@ -1,8 +1,8 @@
 #include <arch/x86/io.h>
-void outb(uint32_t port,uint8_t data) {
+void outb(uint16_t port,uint8_t data) {
     asm("outb %b0, %w1": :"a"(data), "Nd"(port));
 }
-uint8_t inb(uint32_t port) {
+uint8_t inb(uint16_t port) {
      uint8_t result;
      asm("inb %w1, %b0": "=a"(result):"Nd"(port));
      return result;
